Move generic test functions to helpers_test.go
