+++ subscribe_test.go
@@ -1,28 +1,11 @@
 package twitchpubsub
 
 import (
-	"encoding/json"
 	"testing"
 	"time"
 
 	qt "github.com/frankban/quicktest"
 )
-
-type outerMessage struct {
-	Data struct {
-		Topic string `json:"topic"`
-		// Message is an escaped json string
-		Message string `json:"message"`
-	} `json:"data"`
-}
-
-func parseOuterMessage(b []byte) (*outerMessage, error) {
-	msg := outerMessage{}
-	if err := json.Unmarshal(b, &msg); err != nil {
-		return nil, err
-	}
-	return &msg, nil
-}
 
 func TestParseSubscribeEvent(t *testing.T) {
 	c := qt.New(t)
